fix: Fix save-map thumbnail generation
Create thumbnail service used by both save map and share panels
