+++ projects/hslayers/src/components/permalink/public-api.ts
@@ -3,3 +3,4 @@ export * from './share-url.service';
 export * from './share.component';
 export * from './share.module';
 export * from './share.service';
+export * from './share-thumbnail.service';
