+++ projects/hslayers/src/components/permalink/share.service.ts
@@ -15,6 +15,7 @@ import {HsShareUrlService} from './share-url.service';
 import {HsStatusManagerService} from '../save-map/status-manager.service';
 import {HsToastService} from '../layout/toast/toast.service';
 import {HsUtilsService} from '../utils/utils.service';
+import {ShareThumbnailService} from './share-thumbnail.service';
 import {getShowInLayerManager} from '../../common/layer-extensions';
 
 @Injectable({
@@ -35,8 +36,6 @@ export class HsShareService {
     abstract: '',
     shareUrl: '',
   };
-  private renderer: Renderer2;
-
   constructor(
     public HsConfig: HsConfig,
     public HsShareUrlService: HsShareUrlService,
@@ -50,9 +49,8 @@ export class HsShareService {
     public HsToastService: HsToastService,
     public HsLogService: HsLogService,
     private HttpClient: HttpClient,
-    rendererFactory: RendererFactory2
+    public ShareThumbnailService: ShareThumbnailService
   ) {
-    this.renderer = rendererFactory.createRenderer(null, null);
     this.HsEventBusService.mainPanelChanges.subscribe(async () => {
       if (this.HsLayoutService.mainpanel == 'permalink') {
         this.HsShareUrlService.statusSaving = true;
@@ -132,12 +130,14 @@ export class HsShareService {
         'postcompose',
         this.HsUtilsService.debounce(
           () => {
-            this.generateThumbnail(
-              this.HsLayoutService.contentWrapper.querySelector(
-                '.hs-permalink-thumbnail'
-              ),
-              false
-            );
+            if (this.HsLayoutService.mainpanel == 'permalink') {
+              this.generateThumbnail(
+                this.HsLayoutService.contentWrapper.querySelector(
+                  '.hs-permalink-thumbnail'
+                ),
+                false
+              );
+            }
           },
           300,
           false,
@@ -285,128 +285,27 @@ export class HsShareService {
    */
   generateThumbnail($element, newRender: boolean): void {
     this.rendered($element, newRender);
-    if (
-      this.HsLayoutService.mainpanel == 'saveMap' ||
-      this.HsLayoutService.mainpanel == 'permalink' ||
-      this.HsLayoutService.mainpanel == 'shareMap'
-    ) {
-      if ($element === null) {
-        return;
-      }
-      $element.setAttribute('crossOrigin', 'Anonymous');
 
-      if (newRender) {
-        this.rendered.bind(this);
-        this.HsMapService.map.once('postcompose', () =>
-          this.rendered($element, newRender)
-        );
-        this.HsMapService.map.renderSync();
-      } else {
-        this.rendered($element, newRender);
-      }
+    if ($element === null) {
+      return;
     }
-  }
+    $element.setAttribute('crossOrigin', 'Anonymous');
 
-  /**
-   * @param canvas
-   * @param width
-   * @param height
-   */
-  setCanvasSize(canvas, width: number, height: number): void {
-    canvas.width = width;
-    canvas.height = height;
-    canvas.style.width = width + 'px';
-    canvas.style.height = height + 'px';
-  }
-
-  /**
-   * @param ctx
-   */
-  setupContext(ctx): void {
-    ctx.mozImageSmoothingEnabled = false;
-    ctx.webkitImageSmoothingEnabled = false;
-    ctx.msImageSmoothingEnabled = false;
-    ctx.imageSmoothingEnabled = false;
+    if (newRender) {
+      this.HsMapService.map.once('postcompose', () =>
+        this.rendered($element, newRender)
+      );
+      this.HsMapService.map.renderSync();
+    } else {
+      this.rendered($element, newRender);
+    }
   }
 
   rendered($element, newRender?): void {
-    if (!$element) {
-      return;
-    }
-    const collectorCanvas = this.renderer.createElement('canvas');
-    const targetCanvas = this.renderer.createElement('canvas');
-    const width = 256,
-      height = 256;
-    const firstCanvas =
-      this.HsMapService.mapElement.querySelector('.ol-layer canvas');
-    this.setCanvasSize(targetCanvas, width, height);
-    this.setCanvasSize(collectorCanvas, firstCanvas.width, firstCanvas.height);
-    const ctxCollector = collectorCanvas.getContext('2d');
-    const ctxTarget = targetCanvas.getContext('2d');
-    this.setupContext(ctxTarget);
-    this.setupContext(ctxCollector);
-    Array.prototype.forEach.call(
-      this.HsMapService.mapElement.querySelectorAll('.ol-layer canvas'),
-      (canvas) => {
-        if (canvas.width > 0) {
-          //console.log('canvas loop', this.isCanvasTainted(canvas), canvas);
-          /* canvases retrieved from mapElement might be already tainted because they can contain
-           * images (i.e. maps) retrieved from another sources without CORS
-           */
-          const opacity = canvas.parentNode.style.opacity;
-          ctxCollector.globalAlpha = opacity === '' ? 1 : Number(opacity);
-          const transform = canvas.style.transform;
-          // Get the transform parameters from the style's transform matrix
-          const matrix = transform
-            .match(/^matrix\(([^\(]*)\)$/)[1]
-            .split(',')
-            .map(Number);
-          // Apply the transform to the export map context
-          CanvasRenderingContext2D.prototype.setTransform.apply(
-            ctxCollector,
-            matrix
-          );
-          ctxCollector.drawImage(canvas, 0, 0);
-        }
-      }
+    this.data.thumbnail = this.ShareThumbnailService.rendered(
+      $element,
+      newRender
     );
-
-    /* Final render pass */
-    ctxTarget.drawImage(
-      collectorCanvas,
-      Math.floor(collectorCanvas.width / 2 - width / 2),
-      Math.floor(collectorCanvas.height / 2 - height / 2),
-      width,
-      height,
-      0,
-      0,
-      width,
-      height
-    );
-    //console.log('image drawn', this.isCanvasTainted(targetCanvas));
-    /**
-     * from now on, the targetCanvas is also tainted because another tainted canvas was used
-     * to draw inside it
-     */
-
-    try {
-      /**
-       * targetCanvas.toDataURL() fires a SecurityError here
-       * see https://developer.mozilla.org/en-US/docs/Web/HTML/CORS_enabled_image#security_and_tainted_canvases
-       */
-      $element.setAttribute('src', targetCanvas.toDataURL('image/png'));
-      this.data.thumbnail = targetCanvas.toDataURL('image/jpeg', 0.85);
-    } catch (e) {
-      console.log('catched, is tainted?', this.isCanvasTainted(targetCanvas));
-      //console.log(targetCanvas);
-      this.HsLogService.warn(e);
-      $element.setAttribute(
-        'src',
-        this.HsUtilsService.getAssetsPath() + 'img/notAvailable.png'
-      );
-    }
-    $element.style.width = width + 'px';
-    $element.style.height = height + 'px';
   }
 
   private isCanvasTainted(canvas: HTMLCanvasElement): boolean {
